created turbine simulation.

+++ src/include/io.h
@@ -9,7 +9,7 @@
 
 typedef struct OUT {
     uint8_t brake;
-    uint8_t vout;
+    uint16_t vout;
     uint8_t leds;
 } output_t;
